# util/mon: introduce `MonitorName` and reduce allocations