--- a/pkg/base/test_server_args.go
+++ b/pkg/base/test_server_args.go
@@ -498,7 +498,7 @@ func DefaultTestTempStorageConfigWithSize(
 	st *cluster.Settings, maxSizeBytes int64,
 ) TempStorageConfig {
 	monitor := mon.NewMonitor(mon.Options{
-		Name:      "in-mem temp storage",
+		Name:      mon.MakeMonitorName("in-mem temp storage"),
 		Res:       mon.DiskResource,
 		Increment: 1024 * 1024,
 		Settings:  st,