#### util/mon: introduce `MonitorName`

The `MonitorName` type is now used for naming memory monitors. This type
includes an optional 8-byte suffix, helping reduce allocations in a hot
path that added the first 8 bytes of a UUID to the monitor name. Now,
those 8 bytes are included in `MonitorName` allocation free. Allocations
to combine those bytes with the name's prefix in a contiguous string are
only incurred if expensive logging is enabled.

Fixes #136628

Release note: None

#### util/uuid: allocate only 8 bytes for `Short`

`(UUID).Short` previously allocated a 36-byte slice, filled it with
characters, and then sliced off all but the first 8 bytes. Now, it only
allocates and fills 8 bytes.

Release note: None