@mgartner 2024-12-03
Some benchmark results:

```
name                                     old time/op    new time/op    delta
EndToEnd/kv-read/vectorize=on/Simple        193µs ± 2%     193µs ± 2%    ~     (p=0.690 n=5+5)
EndToEnd/kv-read/vectorize=on/Prepared      132µs ± 1%     131µs ± 1%    ~     (p=0.421 n=5+5)
EndToEnd/kv-read/vectorize=off/Simple       192µs ± 1%     193µs ± 1%    ~     (p=0.548 n=5+5)
EndToEnd/kv-read/vectorize=off/Prepared     128µs ± 1%     128µs ± 7%    ~     (p=0.151 n=5+5)

name                                     old alloc/op   new alloc/op   delta
EndToEnd/kv-read/vectorize=on/Simple       36.5kB ± 6%    36.4kB ± 6%    ~     (p=0.548 n=5+5)
EndToEnd/kv-read/vectorize=on/Prepared     25.3kB ± 8%    24.5kB ± 0%    ~     (p=0.206 n=5+4)
EndToEnd/kv-read/vectorize=off/Simple      36.8kB ± 0%    36.7kB ± 0%    ~     (p=0.714 n=5+4)
EndToEnd/kv-read/vectorize=off/Prepared    26.6kB ± 6%    26.2kB ± 4%    ~     (p=0.421 n=5+5)

name                                     old allocs/op  new allocs/op  delta
EndToEnd/kv-read/vectorize=on/Simple          312 ± 4%       310 ± 4%    ~     (p=0.278 n=5+5)
EndToEnd/kv-read/vectorize=on/Prepared        225 ± 7%       218 ± 0%  -2.91%  (p=0.000 n=5+4)
EndToEnd/kv-read/vectorize=off/Simple         300 ± 0%       298 ± 0%  -0.83%  (p=0.029 n=4+4)
EndToEnd/kv-read/vectorize=off/Prepared       212 ± 0%       211 ± 0%  -0.71%  (p=0.029 n=4+4)
```

```
name                                        old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true      80.5µs ± 1%    81.2µs ± 1%  +0.92%  (p=0.032 n=5+5)
FlowSetup/vectorize=true/distribute=false     72.8µs ± 1%    72.9µs ± 2%    ~     (p=0.841 n=5+5)
FlowSetup/vectorize=false/distribute=true     73.8µs ± 2%    74.9µs ± 3%    ~     (p=0.310 n=5+5)
FlowSetup/vectorize=false/distribute=false    66.5µs ± 1%    67.1µs ± 2%    ~     (p=0.421 n=5+5)

name                                        old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true      18.8kB ± 1%    18.8kB ± 1%    ~     (p=0.690 n=5+5)
FlowSetup/vectorize=true/distribute=false     15.6kB ± 0%    15.6kB ± 1%    ~     (p=0.548 n=5+5)
FlowSetup/vectorize=false/distribute=true     25.8kB ± 4%    25.4kB ± 0%  -1.24%  (p=0.040 n=5+5)
FlowSetup/vectorize=false/distribute=false    22.9kB ± 5%    22.9kB ± 6%    ~     (p=0.508 n=5+5)

name                                        old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true         188 ± 0%       186 ± 0%  -1.06%  (p=0.029 n=4+4)
FlowSetup/vectorize=true/distribute=false        163 ± 0%       161 ± 0%  -1.23%  (p=0.000 n=5+4)
FlowSetup/vectorize=false/distribute=true        189 ± 3%       184 ± 0%  -2.44%  (p=0.008 n=5+5)
FlowSetup/vectorize=false/distribute=false       161 ± 0%       159 ± 0%  -1.24%  (p=0.029 n=4+4)
```